fix 相机bug，三角形Material空值，try rayCasting

+++ montecarlo_ray_trace/ray/IntersectResult.h
@@ -8,7 +8,7 @@ class IntersectResult {
 public:
 	Triangle* triangle;
 	bool isIntersect = false;             // 相交
-	double distance = INF;         // 与交点的距离
+	float distance = INF;         // 与交点的距离
 	glm::vec3 intersectPoint = glm::vec3(0, 0, 0);  // 光线交点
 	Material material;              // 交点的表面材质
 	IntersectResult() {};
